Gyroscope Code Updated
Connect gyroscope to SDA and SCL pins (20, 21) , Interrupt on Pin 16.
Functions are turnAngle("left", [degrees of turn], [speed])  Left is the
only direction I have written for yet.  Speed during the turn will
decrease as the turn is completed.  This code has not been tested on a
robot yet.  Enter funtions in the Intructor tab.